Feature/alvaro

+++ backend/src/main/java/com/dando_la_vez/backend/model/Puesto.java
@@ -8,7 +8,6 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.util.List;
-import java.util.Set;
 
 @Entity
 @Table(name = "puestos")
@@ -19,7 +18,7 @@ import java.util.Set;
 public class Puesto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     @Column
     private String nombre;
     @Column
