+++ backend/src/main/java/com/dando_la_vez/backend/repository/PuestoRepository.java
@@ -3,5 +3,5 @@ package com.dando_la_vez.backend.repository;
 import com.dando_la_vez.backend.model.Puesto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PuestoRepository extends JpaRepository<Puesto, Integer> {
+public interface PuestoRepository extends JpaRepository<Puesto, Long> {
 }
